norm: separate parse_pipeline*() functions
